streamlit/requirements.txt
pandas
streamlit
st-dynamodb-connection